refactor: extract leaf presentational components from App.tsx (move-only)
Move the three leaf UI pieces and the Feedback type out of App.tsx into
src/components/, behind the existing barrel:

- SpeakButton   - SpeechSynthesis read-aloud button
- ExamPrompt    - prompt + answer-leak guard + vocab row
- FeedbackPanel - post-answer correct/incorrect/revealed panel
- types.ts      - the Feedback union (shared by App + FeedbackPanel)

No behavioural change: pure move + import rewiring. Also drops four now-
unused lucide icons (AlertTriangle, CheckCircle2, Volume2, XCircle) and a
stale RulesPanel comment left over from the earlier panel extraction.

App.tsx: 873 -> 803 lines. Prereq for the upcoming ChallengePanel + hooks
split.
