+++ src/components/ExamPrompt.tsx
@@ -1,0 +1,53 @@
+import { GraduationCap } from "lucide-react";
+import { type Language } from "../i18n";
+import type { PracticeQuestion } from "../domain/types";
+import { SpeakButton } from "./SpeakButton";
+
+// Renders an exam/cloze/pattern question's prompt: instruction, the
+// sentence with the blank, a neutral pre-answer hint, and (when safe) a
+// surface・reading・meaning vocab row.
+export function ExamPrompt({ question, language }: { question: PracticeQuestion; language: Language }) {
+  // Sentence-pattern items use placeholder surface/reading (the pattern
+  // id) which would render as a meaningless "te-kudasai・te-kudasai・..."
+  // line. Skip the vocab row for those items -- the prompt label
+  // already names the pattern.
+  const isSentencePattern = question.vocabulary.tags?.includes("sentence_pattern");
+  // ANSWER-LEAK GUARD: for 文法形式選擇 items the surface IS the answer
+  // (e.g. surface「ものの」== the correct choice); for 漢字読み items the
+  // reading IS the answer (e.g. reading「とどこおって」). Rendering the
+  // surface・reading・meaning row pre-answer therefore hands the learner
+  // the answer. Only show that row when neither surface nor reading is
+  // one of the expected answers -- which keeps it for cloze items
+  // (待つ -> answer 待って, surface is a legit "which verb" hint) but
+  // hides it for grammar / kanji-reading. The full answer always shows
+  // post-answer in FeedbackPanel.
+  const answerSet = new Set(question.expectedAnswers);
+  const vocabRowLeaksAnswer =
+    answerSet.has(question.vocabulary.surface) || answerSet.has(question.vocabulary.reading);
+  const showVocabRow = !isSentencePattern && !vocabRowLeaksAnswer;
+  // Pre-answer Chinese: prefer the neutral hint when authored; fall back
+  // to the full translation for items that haven't been audited yet
+  // (legacy exam items). The full translation still appears in the
+  // FeedbackPanel post-answer via vocabulary.examples[0].meaningZh.
+  const preAnswerHint = question.hintZh ?? question.promptContextZh;
+  return (
+    <>
+      <p className="word-kind">
+        <GraduationCap aria-hidden="true" />
+        {question.instructionZh}
+      </p>
+      <p className="exam-prompt">
+        {question.promptText}
+        {question.promptText ? (
+          <SpeakButton text={question.promptText} language={language} />
+        ) : null}
+      </p>
+      {preAnswerHint ? <p className="meaning">{preAnswerHint}</p> : null}
+      {showVocabRow ? (
+        <p className="reading">
+          {question.vocabulary.surface}・{question.vocabulary.reading}・{question.vocabulary.meaningZh}
+        </p>
+      ) : null}
+    </>
+  );
+}
