+++ src/components/FeedbackPanel.tsx
@@ -1,0 +1,36 @@
+import { CheckCircle2, XCircle } from "lucide-react";
+import { copy, type Language } from "../i18n";
+import type { Feedback } from "./types";
+
+// Post-answer panel: shows correct/incorrect/revealed status, the
+// accepted answer(s), the explanation, and an example sentence.
+export function FeedbackPanel({
+  feedback,
+  language
+}: {
+  feedback: NonNullable<Feedback>;
+  language: Language;
+}) {
+  const t = copy[language];
+  const isCorrect = feedback.status === "correct";
+  const isRevealed = feedback.status === "revealed";
+  const title = isCorrect ? t.correct : isRevealed ? t.revealed : t.incorrect;
+  const Icon = isCorrect ? CheckCircle2 : XCircle;
+
+  return (
+    <section className={`feedback ${isCorrect ? "correct" : isRevealed ? "revealed" : "incorrect"}`} aria-live="polite">
+      <div className="feedback-title">
+        <Icon aria-hidden="true" />
+        <h2>{title}</h2>
+      </div>
+      <p className="answer-key">{t.answerKey}：{feedback.question.expectedAnswers.join(" / ")}</p>
+      <p>{feedback.question.explanation}</p>
+      {feedback.question.vocabulary.examples[0] ? (
+        <p className="example">
+          {feedback.question.vocabulary.examples[0].japanese}
+          <span>{feedback.question.vocabulary.examples[0].meaningZh}</span>
+        </p>
+      ) : null}
+    </section>
+  );
+}
