+++ src/components/index.ts
@@ -1,6 +1,10 @@
-// Barrel for the top-level view panels extracted from App.tsx.
-// Keeps App's import list to a single line; add future panels here.
+// Barrel for the view panels + leaf presentational components extracted
+// from App.tsx. Keeps App's import list short; add future components here.
 export { HomePanel } from "./HomePanel";
 export { LearningPanel } from "./LearningPanel";
 export { RulesPanel } from "./RulesPanel";
 export { MockExamPanel } from "./MockExamPanel";
+export { ExamPrompt } from "./ExamPrompt";
+export { FeedbackPanel } from "./FeedbackPanel";
+export { SpeakButton } from "./SpeakButton";
+export type { Feedback } from "./types";
